Adding FungibleDice class
Fungible dice are interchangeable.

Examples of fungible dice:
- 3d6
- 4d8 + 2d8

Examples of non-fungible dice:
- 2d4 – 1d4
- 1d12 + 1d8
